Integración de React, React Router (funcion de la naves o cohetes que redirijan a su respectivo espacio) y Three.js con Vite
