+++ index.html
@@ -8,12 +8,10 @@
     <link rel="stylesheet" href="/src/styles.css">
 </head>
 <body>
-    <!-- Cargar Three.js para renderizar el planeta 3D -->
-    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.min.js"></script> -->
-    <!-- Enlace al archivo de JavaScript personalizado -->
-    <div class="section">Bienvenido</div>
-    <div class="section">Aquí están las naves</div>
-    <script type="module" src="/src/main.js"></script> <!-- Enlaza al archivo JS con Three.js -->
-
+    <!-- Este es el contenedor donde React montará la aplicación -->
+    <div id="root"></div>
+    <!-- Cargar el archivo de entrada de React -->
+    <script type="module" src="/src/main.jsx"></script>
 </body>
 </html>
+
