+++ package.json
@@ -14,7 +14,11 @@
     "vite": "^5.4.11"
   },
   "dependencies": {
+    "@vitejs/plugin-react": "^4.3.4",
     "planet-three": "file:",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.1.0",
     "three": "^0.128.0"
   }
 }
