+++ src/App.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import Home from './pages/Home';
+import Ninos from './pages/Ninos';
+import Jovenes from './pages/Jovenes';
+import Padres from './pages/Padres';
+
+function App() {
+    return (
+        <Router>
+            <Routes>
+                <Route path="/" element={<Home />} />
+                <Route path="/ninos" element={<Ninos />} />
+                <Route path="/jovenes" element={<Jovenes />} />
+                <Route path="/padres" element={<Padres />} />
+            </Routes>
+        </Router>
+    );
+}
+
+export default App;
