+++ src/pages/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react'; // Asegúrate de incluir esto
+import ThreeScene from '../components/ThreeScene';
+
+export default function Home() {
+    return (
+        <div>
+            <h1>Explora el Espacio</h1>
+            <ThreeScene />
+        </div>
+    );
+}
