+++ src/pages/Jovenes.jsx
@@ -1,0 +1,3 @@
+export default function Jovenes() {
+    return <h1>Bienvenidos a la sección de Jovenes</h1>;
+}
