+++ src/pages/Ninos.jsx
@@ -1,0 +1,5 @@
+import React from 'react'; // Asegúrate de incluir esto
+
+export default function Ninos() {
+    return <h1>Bienvenidos a la sección de Niños</h1>;
+}
