+++ src/pages/Padres.jsx
@@ -1,0 +1,3 @@
+export default function Padres() {
+    return <h1>Bienvenidos a la sección de Padres</h1>;
+}
